requirements-dev.txt
# Developer tooling aligned with Tessrax governance.
pytest-cov==4.1.0
